% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mlbewith.R
\name{mlbewith}
\alias{mlbewith}
\title{make the descriptions of the model (between/within level)}
\usage{
mlbewith(rowmat, omega3)
}
\arguments{
\item{rowmat}{data matrix for MLSEM}

\item{omega3}{omega for adjusting between-level variance-covariance}
}
\value{
the descriptions of the model (bidirection) in interactive way
}
\description{
make the descriptions of the model (between/within level)
}

