% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mlbi.R
\name{mlbi}
\alias{mlbi}
\title{make the descritpions of the model (bidirection) in interactive way}
\usage{
mlbi(rowmat)
}
\arguments{
\item{rowmat}{data matrix for MLSEM}
}
\value{
the descriptions of the model (bidirection) in interactive way
}
\description{
make the descritpions of the model (bidirection) in interactive way
}

