% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mlcor.R
\name{mlcor}
\alias{mlcor}
\title{compute between.within level correlation}
\usage{
mlcor(x, y, g, fig = F)
}
\arguments{
\item{x}{variable for correlation analysis (1)}

\item{y}{variable for correlation analysis (2)}

\item{g}{grouping variable}

\item{fig}{if fig=TRUE, mlcor draws the scatterplot}
}
\value{
compute between/within level correlation
}
\description{
compute between.within level correlation
}

