% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mlcori.R
\name{mlcori}
\alias{mlcori}
\title{compute between/within level correlation in interactive way}
\usage{
mlcori(rowmat)
}
\arguments{
\item{rowmat}{data matrix for correlation analysis}
}
\value{
compute between/within level correlation in interactive way
}
\description{
compute between/within level correlation in interactive way
}

