% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mlcovar.R
\name{mlcovar}
\alias{mlcovar}
\title{compute between/within level variance-covariance matrix}
\usage{
mlcovar(rowmat)
}
\arguments{
\item{rowmat}{data matrix for computing variance-covariance matrix. the first column must be grouping variable.}
}
\value{
compute between/within level variance-covariance matrix
}
\description{
compute between/within level variance-covariance matrix
}

