% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mlcovari.R
\name{mlcovari}
\alias{mlcovari}
\title{compute between/within level variance-covariance matrix in interactive way.}
\usage{
mlcovari(rowmat)
}
\arguments{
\item{rowmat}{data matrix for computing variance-covariance matrix. the first column must be grouping variable.}
}
\value{
compute between/within level variance-covariance matrix in interactive way.
}
\description{
compute between/within level variance-covariance matrix in interactive way.
}

