% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mllavar.R
\name{mllavar}
\alias{mllavar}
\title{conduct MLSEM by using lavaan function (in interactive way)}
\usage{
mllavar(rowmat, sub = NULL, unimodel = NULL, bimodel = NULL,
  fig = FALSE, verbose = FALSE)
}
\arguments{
\item{rowmat}{data matrix for MLSEM. the first column must be grouping variable.}

\item{sub}{if sub=TRUE, you can choose variables for MLSEM from data matrix in interactive way. If sub=FALSE, all the variables in rowmat were analyzed.}

\item{unimodel}{desctiptions of model (unidirection). if unimodel=NULL, you can make descriptions of model (Unidirection) in interactive way.}

\item{bimodel}{desctiptions of model (bidirection). if bimodel=NULL, you can make descriptions of model (bidirection) in interactive way.}

\item{fig}{if fig=TRUE, draw the figure of results by using semPaths function}

\item{verbose}{if verbose=TRUE, show the process of iterations.}
}
\value{
conduct MLSEM by using lavaan function (in interactive way)
}
\description{
conduct MLSEM by using lavaan function (in interactive way)
}

