% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mllavartest.R
\docType{data}
\name{mllavartest}
\alias{mllavartest}
\title{data example for MLLaVar}
\format{A data frame with 163 rows and 5 variables}
\usage{
data(mllavartest)
}
\description{
The variables are follows:
}
\details{
\itemize{
  \item group group variable
  \item time independent variable (between level)
  \item collectivism independent variable 1
  \item individualism independent variable 2
  \item satisfaction dependent variable
}
}
\keyword{datasets}

