% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mlsubset.R
\name{mlsubset}
\alias{mlsubset}
\title{make data matrix for mlcovar function in interactive way}
\usage{
mlsubset(rowmat)
}
\arguments{
\item{rowmat}{data matrix for reshaping}
}
\value{
data matrix for mlcovar function
}
\description{
make data matrix for mlcovar function in interactive way
}

