% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mluni.R
\name{mluni}
\alias{mluni}
\title{make the descriptions of the model (unidirection) in interactive way}
\usage{
mluni(rowmat)
}
\arguments{
\item{rowmat}{data matrix for MLSEM}
}
\value{
the descriptions of the model (unidirection) in interactive way
}
\description{
make the descriptions of the model (unidirection) in interactive way
}

